More hardcoding cleaning with inf3
-Add skill lists for :
--spell blocked by statis
--spell blocked by kagehumi
--spell range affected by
AC_VULTURE|GS_SNAKEEYE|NJ_SHADOWJUMP|WL_RADIUSRA_RESEARCHTRAP
--spell that can't be copied
--spell that can be using while riding warg|mado
-Add a boolean return value for delequip (script cmd)
-Fix monster (script cmd) hardcoded cap for ai. (was preventing to put
ai_faw)

+++ src/char/inter.c
@@ -667,7 +667,7 @@ static int inter_config_read(const char* cfgName)
 			strcpy(default_codepage,w2);
 		}
 		else if(!strcmpi(w1,"party_share_level"))
-			party_share_level = atoi(w2);
+			party_share_level = (unsigned int)atof(w2);
 		else if(!strcmpi(w1,"log_inter"))
 			log_inter = atoi(w2);
 		else if(!strcmpi(w1,"import"))
